Use a chip-wide reset. Does something now.
Did nothing in the chip or in ISE Simulator. Now it does something
in both, albeit wrong. Datapaths cannot exceed 32 bits. The delta
sigma DAC is doing something ridiculous, don't know why. Actually
it looks more like an interpolator issue. But it's activity visible
on the scope, that's enough for tonight.